Refactor for better organisation
